Move the PdfPageEventForwarder and rewrite the notification code
